+++ repos/iiif-storage/src/Media/CrowdSourcingBannerIngester.php
@@ -73,6 +73,7 @@ class CrowdSourcingBannerIngester extends AbstractIngester implements IngesterIn
                     'label' => 'Collection ID', // @translate
                     'info' => 'ID Main collection used in call to action, defaults to all collections in site (/collections/all)', // @translate
                 ]),
+            $this->getLocaleField(),
         ];
     }