+++ repos/iiif-storage/src/Media/LatestAnnotatedImagesRenderer.php
@@ -2,6 +2,7 @@
 
 namespace IIIFStorage\Media;
 
+use Digirati\OmekaShared\Helper\LocaleHelper;
 use Doctrine\DBAL\Connection;
 use IIIFStorage\JsonBuilder\CanvasBuilder;
 use IIIFStorage\Repository\CanvasRepository;
@@ -13,7 +14,7 @@ use Zend\View\Model\ViewModel;
 use Zend\View\Renderer\PhpRenderer;
 use ZfcTwig\View\TwigRenderer;
 
-class LatestAnnotatedImagesRenderer implements RendererInterface, MediaPageBlockDualRender
+class LatestAnnotatedImagesRenderer implements RendererInterface, MediaPageBlockDualRender, LocalisedMedia
 {
     use RenderMedia;
 
@@ -37,19 +38,25 @@ class LatestAnnotatedImagesRenderer implements RendererInterface, MediaPageBlock
      * @var Router
      */
     private $router;
+    /**
+     * @var LocaleHelper
+     */
+    private $localeHelper;
 
     public function __construct(
         TwigRenderer $twig,
         Connection $connection,
         CanvasRepository $repository,
         CanvasBuilder $canvasBuilder,
-        Router $router
+        Router $router,
+        LocaleHelper $localeHelper
     ) {
         $this->twig = $twig;
         $this->connection = $connection;
         $this->repository = $repository;
         $this->canvasBuilder = $canvasBuilder;
         $this->router = $router;
+        $this->localeHelper = $localeHelper;
     }
 
     public function renderFromData(PhpRenderer $view, array $data, array $options = [])
@@ -96,5 +103,10 @@ class LatestAnnotatedImagesRenderer implements RendererInterface, MediaPageBlock
     public function pageBlockOptions(SitePageBlockRepresentation $pageBlock): array
     {
         return [];
+    }
+
+    public function getLang(): string
+    {
+        return $this->localeHelper->getLocale();
     }
 }